% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_halved.R
\name{halved}
\alias{halved}
\title{Create a function to return the halved value of numbers}
\usage{
halved(x)
}
\arguments{
\item{x}{Input vector, must be a character vector}
}
\value{
a character vector that is the halved value of input vector x
}
\description{
Descriptions about the function
}
