% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_halved.R
\name{squared}
\alias{squared}
\title{Create a function to return the squares of numbers}
\usage{
squared(x)
}
\arguments{
\item{x}{Input vector, must be a character vector}
}
\value{
a character vector that is the squared value of input vector x
}
\description{
Descriptions about the function
}
